+++ pkg/operator/operatorclient/interfaces.go
@@ -1,0 +1,9 @@
+package operatorclient
+
+const (
+	EtcdNamespaceName                     = "kube-system"
+	UserSpecifiedGlobalConfigNamespace    = "openshift-config"
+	MachineSpecifiedGlobalConfigNamespace = "openshift-config-managed"
+	OperatorNamespace                     = "openshift-kube-apiserver-operator"
+	TargetNamespaceName                   = "openshift-kube-apiserver"
+)
