--- pkg/operator/target_config_reconciler.go
@@ -1,213 +1,0 @@
-package operator
-
-import (
-	"fmt"
-	"reflect"
-	"time"
-
-	"github.com/golang/glog"
-
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	"k8s.io/apimachinery/pkg/util/sets"
-	"k8s.io/apimachinery/pkg/util/wait"
-	"k8s.io/client-go/informers"
-	"k8s.io/client-go/kubernetes"
-	corev1listers "k8s.io/client-go/listers/core/v1"
-	"k8s.io/client-go/tools/cache"
-	"k8s.io/client-go/util/workqueue"
-
-	operatorv1 "github.com/openshift/api/operator/v1"
-	operatorconfigclientv1alpha1 "github.com/openshift/cluster-kube-apiserver-operator/pkg/generated/clientset/versioned/typed/kubeapiserver/v1alpha1"
-	operatorconfiginformerv1alpha1 "github.com/openshift/cluster-kube-apiserver-operator/pkg/generated/informers/externalversions/kubeapiserver/v1alpha1"
-	"github.com/openshift/library-go/pkg/operator/events"
-	"github.com/openshift/library-go/pkg/operator/v1helpers"
-)
-
-type TargetConfigReconciler struct {
-	targetImagePullSpec string
-
-	operatorConfigClient operatorconfigclientv1alpha1.KubeapiserverV1alpha1Interface
-
-	kubeClient                          kubernetes.Interface
-	configMapListerForOperatorNamespace corev1listers.ConfigMapLister
-	eventRecorder                       events.Recorder
-
-	// queue only ever has one item, but it has nice error handling backoff/retry semantics
-	queue workqueue.RateLimitingInterface
-}
-
-func NewTargetConfigReconciler(
-	targetImagePullSpec string,
-	operatorConfigInformer operatorconfiginformerv1alpha1.KubeAPIServerOperatorConfigInformer,
-	kubeInformersForOpenshiftKubeAPIServerNamespace informers.SharedInformerFactory,
-	kubeInformersForOperatorNamespace informers.SharedInformerFactory,
-	operatorConfigClient operatorconfigclientv1alpha1.KubeapiserverV1alpha1Interface,
-	kubeClient kubernetes.Interface,
-	eventRecorder events.Recorder,
-) *TargetConfigReconciler {
-	c := &TargetConfigReconciler{
-		targetImagePullSpec: targetImagePullSpec,
-
-		operatorConfigClient:                operatorConfigClient,
-		kubeClient:                          kubeClient,
-		configMapListerForOperatorNamespace: kubeInformersForOperatorNamespace.Core().V1().ConfigMaps().Lister(),
-		eventRecorder:                       eventRecorder,
-
-		queue: workqueue.NewNamedRateLimitingQueue(workqueue.DefaultControllerRateLimiter(), "TargetConfigReconciler"),
-	}
-
-	operatorConfigInformer.Informer().AddEventHandler(c.eventHandler())
-	kubeInformersForOpenshiftKubeAPIServerNamespace.Rbac().V1().Roles().Informer().AddEventHandler(c.eventHandler())
-	kubeInformersForOpenshiftKubeAPIServerNamespace.Rbac().V1().RoleBindings().Informer().AddEventHandler(c.eventHandler())
-	kubeInformersForOpenshiftKubeAPIServerNamespace.Core().V1().ConfigMaps().Informer().AddEventHandler(c.eventHandler())
-	kubeInformersForOpenshiftKubeAPIServerNamespace.Core().V1().Secrets().Informer().AddEventHandler(c.eventHandler())
-	kubeInformersForOpenshiftKubeAPIServerNamespace.Core().V1().ServiceAccounts().Informer().AddEventHandler(c.eventHandler())
-	kubeInformersForOpenshiftKubeAPIServerNamespace.Core().V1().Services().Informer().AddEventHandler(c.eventHandler())
-
-	// we react to some config changes
-	kubeInformersForOperatorNamespace.Core().V1().ConfigMaps().Informer().AddEventHandler(c.eventHandler())
-
-	// we only watch some namespaces
-	kubeInformersForOpenshiftKubeAPIServerNamespace.Core().V1().Namespaces().Informer().AddEventHandler(c.namespaceEventHandler())
-
-	return c
-}
-
-func (c TargetConfigReconciler) sync() error {
-	operatorConfig, err := c.operatorConfigClient.KubeAPIServerOperatorConfigs().Get("instance", metav1.GetOptions{})
-	if err != nil {
-		return err
-	}
-
-	operatorConfigOriginal := operatorConfig.DeepCopy()
-
-	switch operatorConfig.Spec.ManagementState {
-	case operatorv1.Unmanaged:
-		return nil
-
-	case operatorv1.Removed:
-		// TODO probably just fail
-		return nil
-	}
-
-	// block until config is obvserved
-	if len(operatorConfig.Spec.ObservedConfig.Raw) == 0 {
-		glog.Info("Waiting for observed configuration to be available")
-		return nil
-	}
-
-	requeue, err := createTargetConfigReconciler_v311_00_to_latest(c, c.eventRecorder, operatorConfig)
-	if requeue && err == nil {
-		return fmt.Errorf("synthetic requeue request")
-	}
-
-	if err != nil {
-		if !reflect.DeepEqual(operatorConfigOriginal, operatorConfig) {
-			v1helpers.SetOperatorCondition(&operatorConfig.Status.Conditions, operatorv1.OperatorCondition{
-				Type:    operatorv1.OperatorStatusTypeFailing,
-				Status:  operatorv1.ConditionTrue,
-				Reason:  "StatusUpdateError",
-				Message: err.Error(),
-			})
-			if _, updateError := c.operatorConfigClient.KubeAPIServerOperatorConfigs().UpdateStatus(operatorConfig); updateError != nil {
-				glog.Error(updateError)
-			}
-		}
-		return err
-	}
-
-	return nil
-}
-
-// Run starts the kube-apiserver and blocks until stopCh is closed.
-func (c *TargetConfigReconciler) Run(workers int, stopCh <-chan struct{}) {
-	defer utilruntime.HandleCrash()
-	defer c.queue.ShutDown()
-
-	glog.Infof("Starting TargetConfigReconciler")
-	defer glog.Infof("Shutting down TargetConfigReconciler")
-
-	// doesn't matter what workers say, only start one.
-	go wait.Until(c.runWorker, time.Second, stopCh)
-
-	<-stopCh
-}
-
-func (c *TargetConfigReconciler) runWorker() {
-	for c.processNextWorkItem() {
-	}
-}
-
-func (c *TargetConfigReconciler) processNextWorkItem() bool {
-	dsKey, quit := c.queue.Get()
-	if quit {
-		return false
-	}
-	defer c.queue.Done(dsKey)
-
-	err := c.sync()
-	if err == nil {
-		c.queue.Forget(dsKey)
-		return true
-	}
-
-	utilruntime.HandleError(fmt.Errorf("%v failed with : %v", dsKey, err))
-	c.queue.AddRateLimited(dsKey)
-
-	return true
-}
-
-// eventHandler queues the operator to check spec and status
-func (c *TargetConfigReconciler) eventHandler() cache.ResourceEventHandler {
-	return cache.ResourceEventHandlerFuncs{
-		AddFunc:    func(obj interface{}) { c.queue.Add(workQueueKey) },
-		UpdateFunc: func(old, new interface{}) { c.queue.Add(workQueueKey) },
-		DeleteFunc: func(obj interface{}) { c.queue.Add(workQueueKey) },
-	}
-}
-
-// this set of namespaces will include things like logging and metrics which are used to drive
-var interestingNamespaces = sets.NewString(targetNamespaceName)
-
-func (c *TargetConfigReconciler) namespaceEventHandler() cache.ResourceEventHandler {
-	return cache.ResourceEventHandlerFuncs{
-		AddFunc: func(obj interface{}) {
-			ns, ok := obj.(*corev1.Namespace)
-			if !ok {
-				c.queue.Add(workQueueKey)
-			}
-			if ns.Name == targetNamespaceName {
-				c.queue.Add(workQueueKey)
-			}
-		},
-		UpdateFunc: func(old, new interface{}) {
-			ns, ok := old.(*corev1.Namespace)
-			if !ok {
-				c.queue.Add(workQueueKey)
-			}
-			if ns.Name == targetNamespaceName {
-				c.queue.Add(workQueueKey)
-			}
-		},
-		DeleteFunc: func(obj interface{}) {
-			ns, ok := obj.(*corev1.Namespace)
-			if !ok {
-				tombstone, ok := obj.(cache.DeletedFinalStateUnknown)
-				if !ok {
-					utilruntime.HandleError(fmt.Errorf("couldn't get object from tombstone %#v", obj))
-					return
-				}
-				ns, ok = tombstone.Obj.(*corev1.Namespace)
-				if !ok {
-					utilruntime.HandleError(fmt.Errorf("tombstone contained object that is not a Namespace %#v", obj))
-					return
-				}
-			}
-			if ns.Name == targetNamespaceName {
-				c.queue.Add(workQueueKey)
-			}
-		},
-	}
-}
