--- pkg/operator/target_config_reconciler_v311_00.go
@@ -1,168 +1,0 @@
-package operator
-
-import (
-	"crypto/x509"
-	"fmt"
-	"reflect"
-
-	corev1 "k8s.io/api/core/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	coreclientv1 "k8s.io/client-go/kubernetes/typed/core/v1"
-	corev1listers "k8s.io/client-go/listers/core/v1"
-	"k8s.io/client-go/util/cert"
-
-	operatorv1 "github.com/openshift/api/operator/v1"
-	"github.com/openshift/cluster-kube-apiserver-operator/pkg/apis/kubeapiserver/v1alpha1"
-	"github.com/openshift/cluster-kube-apiserver-operator/pkg/operator/v311_00_assets"
-	"github.com/openshift/cluster-kube-apiserver-operator/pkg/version"
-	"github.com/openshift/library-go/pkg/crypto"
-	"github.com/openshift/library-go/pkg/operator/events"
-	"github.com/openshift/library-go/pkg/operator/resource/resourceapply"
-	"github.com/openshift/library-go/pkg/operator/resource/resourcemerge"
-	"github.com/openshift/library-go/pkg/operator/resource/resourceread"
-	"github.com/openshift/library-go/pkg/operator/resourcesynccontroller"
-	"github.com/openshift/library-go/pkg/operator/v1helpers"
-)
-
-// createTargetConfigReconciler_v311_00_to_latest takes care of creation of valid resources in a fixed name.  These are inputs to other control loops.
-// returns whether or not requeue and if an error happened when updating status.  Normally it updates status itself.
-func createTargetConfigReconciler_v311_00_to_latest(c TargetConfigReconciler, recorder events.Recorder, operatorConfig *v1alpha1.KubeAPIServerOperatorConfig) (bool, error) {
-	operatorConfigOriginal := operatorConfig.DeepCopy()
-	errors := []error{}
-
-	directResourceResults := resourceapply.ApplyDirectly(c.kubeClient, c.eventRecorder, v311_00_assets.Asset,
-		"v3.11.0/kube-apiserver/ns.yaml",
-		"v3.11.0/kube-apiserver/svc.yaml",
-	)
-
-	for _, currResult := range directResourceResults {
-		if currResult.Error != nil {
-			errors = append(errors, fmt.Errorf("%q (%T): %v", currResult.File, currResult.Type, currResult.Error))
-		}
-	}
-
-	_, _, err := manageKubeApiserverConfig(c.kubeClient.CoreV1(), recorder, operatorConfig)
-	if err != nil {
-		errors = append(errors, fmt.Errorf("%q: %v", "configmap/config", err))
-	}
-	_, _, err = managePod(c.kubeClient.CoreV1(), recorder, operatorConfig, c.targetImagePullSpec)
-	if err != nil {
-		errors = append(errors, fmt.Errorf("%q: %v", "configmap/kube-apiserver-pod", err))
-	}
-	_, _, err = manageClientCABundle(c.configMapListerForOperatorNamespace, c.kubeClient.CoreV1(), recorder)
-	if err != nil {
-		errors = append(errors, fmt.Errorf("%q: %v", "configmap/client-ca", err))
-	}
-
-	if len(errors) > 0 {
-		v1helpers.SetOperatorCondition(&operatorConfig.Status.Conditions, operatorv1.OperatorCondition{
-			Type:    "TargetConfigReconcilerFailing",
-			Status:  operatorv1.ConditionTrue,
-			Reason:  "SynchronizationError",
-			Message: v1helpers.NewMultiLineAggregate(errors).Error(),
-		})
-		if !reflect.DeepEqual(operatorConfigOriginal, operatorConfig) {
-			_, updateError := c.operatorConfigClient.KubeAPIServerOperatorConfigs().UpdateStatus(operatorConfig)
-			return true, updateError
-		}
-		return true, nil
-	}
-
-	v1helpers.SetOperatorCondition(&operatorConfig.Status.Conditions, operatorv1.OperatorCondition{
-		Type:   "TargetConfigReconcilerFailing",
-		Status: operatorv1.ConditionFalse,
-	})
-	if !reflect.DeepEqual(operatorConfigOriginal, operatorConfig) {
-		_, updateError := c.operatorConfigClient.KubeAPIServerOperatorConfigs().UpdateStatus(operatorConfig)
-		if updateError != nil {
-			return true, updateError
-		}
-	}
-
-	return false, nil
-}
-
-func manageKubeApiserverConfig(client coreclientv1.ConfigMapsGetter, recorder events.Recorder, operatorConfig *v1alpha1.KubeAPIServerOperatorConfig) (*corev1.ConfigMap, bool, error) {
-	configMap := resourceread.ReadConfigMapV1OrDie(v311_00_assets.MustAsset("v3.11.0/kube-apiserver/cm.yaml"))
-	defaultConfig := v311_00_assets.MustAsset("v3.11.0/kube-apiserver/defaultconfig.yaml")
-	requiredConfigMap, _, err := resourcemerge.MergeConfigMap(configMap, "config.yaml", nil, defaultConfig, operatorConfig.Spec.ObservedConfig.Raw, operatorConfig.Spec.UnsupportedConfigOverrides.Raw)
-	if err != nil {
-		return nil, false, err
-	}
-	return resourceapply.ApplyConfigMap(client, recorder, requiredConfigMap)
-}
-
-func managePod(client coreclientv1.ConfigMapsGetter, recorder events.Recorder, operatorConfig *v1alpha1.KubeAPIServerOperatorConfig, imagePullSpec string) (*corev1.ConfigMap, bool, error) {
-	required := resourceread.ReadPodV1OrDie(v311_00_assets.MustAsset("v3.11.0/kube-apiserver/pod.yaml"))
-	required.Spec.Containers[0].ImagePullPolicy = corev1.PullAlways
-	if len(imagePullSpec) > 0 {
-		required.Spec.Containers[0].Image = imagePullSpec
-	}
-	required.Spec.Containers[0].Args = append(required.Spec.Containers[0].Args, fmt.Sprintf("-v=%d", 2))
-
-	configMap := resourceread.ReadConfigMapV1OrDie(v311_00_assets.MustAsset("v3.11.0/kube-apiserver/pod-cm.yaml"))
-	configMap.Data["pod.yaml"] = resourceread.WritePodV1OrDie(required)
-	configMap.Data["forceRedeploymentReason"] = operatorConfig.Spec.ForceRedeploymentReason
-	configMap.Data["version"] = version.Get().String()
-	return resourceapply.ApplyConfigMap(client, recorder, configMap)
-}
-
-func manageClientCABundle(lister corev1listers.ConfigMapLister, client coreclientv1.ConfigMapsGetter, recorder events.Recorder) (*corev1.ConfigMap, bool, error) {
-	inputs := []resourcesynccontroller.ResourceLocation{
-		// this is from the installer and contains the value they think we should have
-		{Namespace: operatorNamespace, Name: "initial-client-ca"},
-		// this is from kube-controller-manager and indicates the ca-bundle.crt to verify their signatures (kubelet client certs)
-		{Namespace: operatorNamespace, Name: "csr-controller-ca"},
-		// this bundle is what this operator uses to mint new client certs it directly manages
-		{Namespace: operatorNamespace, Name: "managed-kube-apiserver-client-ca-bundle"},
-	}
-
-	certificates := []*x509.Certificate{}
-	for _, input := range inputs {
-		inputConfigMap, err := lister.ConfigMaps(input.Namespace).Get(input.Name)
-		if apierrors.IsNotFound(err) {
-			continue
-		}
-		if err != nil {
-			return nil, false, err
-		}
-
-		inputContent := inputConfigMap.Data["ca-bundle.crt"]
-		if len(inputContent) == 0 {
-			continue
-		}
-		inputCerts, err := cert.ParseCertsPEM([]byte(inputContent))
-		if err != nil {
-			return nil, false, fmt.Errorf("configmap/%s in %q is malformed: %v", input.Name, input.Namespace, err)
-		}
-		certificates = append(certificates, inputCerts...)
-	}
-
-	finalCertificates := []*x509.Certificate{}
-	// now check for duplicates. n^2, but super simple
-	for i := range certificates {
-		found := false
-		for j := range finalCertificates {
-			if reflect.DeepEqual(certificates[i].Raw, finalCertificates[j].Raw) {
-				found = true
-				break
-			}
-		}
-		if !found {
-			finalCertificates = append(finalCertificates, certificates[i])
-		}
-	}
-	caBytes, err := crypto.EncodeCertificates(finalCertificates...)
-	if err != nil {
-		return nil, false, err
-	}
-
-	requiredConfigMap := &corev1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{Namespace: targetNamespaceName, Name: "client-ca"},
-		Data: map[string]string{
-			"ca-bundle.crt": string(caBytes),
-		},
-	}
-	return resourceapply.ApplyConfigMap(client, recorder, requiredConfigMap)
-}
